fix: consolidate executor-review reconcile files for pkg-file-count
